+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.0
 
 * Internal representation of tags changed meaning all previous tags won't work with this version
+* Change `--recursive` global parameter to `--max-depth` [#19](https://github.com/wojciechkepka/wutag/pull/19)
 
 # 0.2.0